+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 7.0.0 (unreleased)
 
 * Have several functions take arguments by value, per clippy's `trivially_copy_pass_by_ref`
+* Remove `addresses()` from `CNameResults` - CNAME queries don't return addresses
 
 ## 6.0.0 (28 May 2018)